Improve handling of 7z archives extraction
Extraction code is no longer defined in the dependencies-checking part
of the library.

When trying to extract the contents of an archive but no extractor is
present (can happen if dependencies checking is faulty), fail with an
explicit error and a link to our bugs tracker.

Silence output of 7zr/7za/unar when handling 7z archives.