Star Wars: Galactic Battlegrounds — Work around a crash on resolution change
The game crashes on resolution change. This happens when starting or
loading a game while the game resolution is set to something else than
800x600, or when trying to change the game resolution during gameplay.

The crash on resolution change is triggered only when
[HKEY_CURRENT_USER\Software\Wine\Direct3D] "renderer" is set to "gl".
When set to "gdi" instead the resolution change happen with no error.

See https://forge.dotslashplay.it/play.it/games/-/issues/464 and
https://bugs.winehq.org/show_bug.cgi?id=49649 for more details.