[arch] Make .MTREE creation optional
Some Arch users complained that creating a .MTREE file took too much
time, because it needs to compute hashes of every file in the archive,
which can be quite long for big games. However, this file is needed by
pacman to ensure integrity of the archive. Hence it has been made
optional, defaulting to creating the file.