Provide icons inclusion using a single function
A single function is now available to game scripts, for icons inclusion: icons_inclusion

USAGE: icons_inclusion $application[…]

This function is the one that should be called from game scripts,
it can take several applications as its arguments,
and default to handle all applications if none are explicitely given.

It relies on APP_xxx_ICONS_LIST / APP_xxx_ICON, like the previous
functions. It looks for the icon files under the following path:
${PLAYIT_WORKDIR}/gamedata/${CONTENT_PATH_DEFAULT}

Compatibility wrappers are available for the following deprecated functions:
- icons_get_from_package
- icons_get_from_workdir