Add ability to request a specific prefix type from the game script
The prefix type can be set for a given application using
APP_xxx_PREFIX_TYPE or for all applications for the current game using
APPLICATIONS_PREFIX_TYPE.

The allowed values are:
- "symlinks", the default, generate our usual symbolic links farm
- "none", no prefix is generated, the game is run from the read-only
  system directory

If no value is set, most application types default to the "symlinks"
prefix type. ScummVM and ResidualVM applications default to "none"
instead.