Launchers - Compute path to game prefix using a dedicated function
The new default prefix path is:
~/.cache/play.it/prefixes/${GAME_ID}

If $XDG_CACHE_HOME is set, it is used instead of ~/.cache

The full prefix path can be overridden for the current session by setting
the new environment variable $PLAYIT_PREFIX_PATH

The undocumented environment variable $PREFIX_ID is no longer supported.

Prefixes migration from the old path (under ~/.local/share/play.it) to
the new one is not done automatically, and no tool is provided to do
such a migration. The old prefix is not deleted either.