New variable to set archive content path: CONTENT_xxx_PATH
CONTENT_xxx_PATH, with "xxx" a content identifier, is a new variable
that game scripts can define to set the inner archive path to some
content that should be fetched. If unset for a given content identifier,
it falls back to the value of CONTENT_PATH_DEFAULT.

ARCHIVE_xxx_PATH is still supported for compatibility with old game
scripts, but will be silently ignored for game scripts targeting
./play.it ≥ 2.19. For game scripts targeting ./play.it = 2.18, both
variables are supported.