New variable to set archive content files: CONTENT_xxx_FILES
CONTENT_xxx_FILES, with "xxx" a content identifier, is a new variable
that game scripts can define to set the list of files that should be
fetched from a given path.

ARCHIVE_xxx_FILES is still supported for compatibility with old game
scripts, but will be silently ignored for game scripts targeting
./play.it ≥ 2.19. For game scripts targeting ./play.it = 2.18, both
variables are supported.