When some library dependencies are not known, display their list before the installation instructions
A list of unknown libraries is displayed when print_instructions is
called. This list relies on a file stored under $PLAYIT_WORKDIR, so this
path should no longer be deleted before the call to print_instructions.
If it is deleted this will not trigger an error, but the warning about
unknown library requirements will not be shown.