Add ability to get the name of a context-specific variable
Prior to this update, fetching a context-specific value could be done
with "get_context_specific_value" but the name of the variable storing
this value could not be found easily.

This update adds a new function "context_specific_name" that prints the
name of the variable the context-specific value is fetched from.