Change egentoo output paths
Instead of making packages in `data`, `amd64` and `x86` directories,
ebuilds are now created in an `overlay/games-playit` directory and
packages in a `packages` directory.

This has the side effect of making similarly named packages - for
instance binary packages targetting different architectures - overwrite
each other. This shall be taken care of in a follow-up commit.