Update function printing the path where a package is prepared
The function is renamed to "package_path".
A compatibility "package_get_path" is provided.

The function now relies on "package_name" to get the name of the
directory where the package is prepared. "egentoo" target package format
is an exception, as the function "egentoo_package_name" returns the same
value for all packages.

Packages are prepared in ${PLAYIT_WORKDIR}/packages
instead of the root of $PLAYIT_WORKDIR.