Add scummvm and residualvm dependency keywords
Since the command dependency system has not been implemented yet, this
relies on the legacy dependency system.

Moreover, although the “residualvm” keyword has been added for
archlinux, such a package does not exist in the official repositories
nor in the AUR. It is up to archlinux users to write a PKGBUILD for this
program.