Provide a new function to get a variable context-specific name
Deprecation warning:

- "context_specific_name" is dropped, "context_name" should be used
  instead. No compatibility alias is provided.

New function:

- context_name
  USAGE: context_name $variable_name
  RETURN: the name of the variable containing the context-specific value,
          or an empty string

The context can no longer be limited to either archive or package, the
name returned is the first one of a variable containing a non-empty
value, with the following context priority order:
- archive-specific
- package-specific
- default (base variable name)
- empty (no name returned)

The new function no longer returns the base variable name if it is the
name of an empty variable. In such cases, it returns an empty string
instead.