Use new functions to set options during initialisation
Deprecation notices

* The option "--icons yes|no" is dropped. "--icons no" is replaced with "--no-icons". There is no replacement for "--icons yes", that is still the default behaviour.
* The option "--skip-free-space-check" is replaced with "--no-free-space-check". This is a drop-in replacement.
* Support for short options is dropped. "--help" can be used instead of "-h", "--version" instead of "-v" and "--config-file" instead of "-c".

Arguments set in the configuration are used to update the default value
of options, not their current value.

The following options can no longer be set from the configuration file:
- help
- list-packages
- list-requirements
- config-file
- the path to the archive

Arguments given on the command line do not override values set from the
user environment.

Support is dropped for --option=help, it was not available reliably anyway.