Set a new list of valid values for "compression" option
The new valid values are:
- none
- speed
- size

The old values are still valid for game scripts targeting ./play.it ≤ 2.22,
but will no longer be available for game scripts targeting ./play.it ≥ 2.23.
Using these old values will trigger the display of a deprecation warning.

Legacy compression options can only be used if "--package" is explicitly set.