Run a late requirements check for icons extraction
A late requirements check can be useful in case it was skipped earlier because of an empty icons list.
This can happen if both the application type and icon are implicit, like with WINE games where the icons source is the game binary.
Failing at the icons extraction step is not ideal because the archive contents extraction already happened, but at least the failure is explicit instead of silent.